Moved unit test to own file
